verbindung zum Event erstellen
Ich habe die Beziehung in der DbContext Klasse zum Event erstellt. außerdem habe ich in der Event Model NAvigationseigenschaft zum User erstellt.
