+++ EventManagmentSystem/DAL/EventDbContext.cs
@@ -1,5 +1,4 @@
-﻿using EventManagmentSystem.Models;
-using Microsoft.EntityFrameworkCore;
+﻿using Microsoft.EntityFrameworkCore;
 using EventManagmentSystem.Models.DbModel;
 
 namespace EventManagmentSystem.DAL
@@ -11,10 +10,7 @@ namespace EventManagmentSystem.DAL
         { }
         public DbSet<User> Users { get; set; }
         public DbSet<Event> Events { get; set; }
-        // public DbSet<Event> Events { get; set; }
-
         //Das brauchen wir später wegen der Beziehung zu der Event-Tabelle und Bookingtable
-       /*
         protected override void OnModelCreating(ModelBuilder modelBuilder)
         {
             modelBuilder.Entity<User>()
@@ -23,6 +19,5 @@ namespace EventManagmentSystem.DAL
                 .HasForeignKey(e => e.UserId)
                 .OnDelete(DeleteBehavior.Cascade);  // Setzt die Löschkaskade
         }
-       */
     }
 }
