--- EventManagmentSystem/Models/EventOverview.cs
@@ -1,6 +1,0 @@
-﻿namespace EventManagmentSystem.Models;
-
-public class EventOverview
-{
-    public List<Event> ListOfEvents { get; set; }
-}
