+++ EventManagmentSystem/Models/ViewModel/EventOverview.cs
@@ -1,0 +1,8 @@
+﻿using EventManagmentSystem.Models.DbModel;
+
+namespace EventManagmentSystem.Models;
+
+public class EventOverview
+{
+    public List<Event> ListOfEvents { get; set; }
+}
